refactor: refactored naming on reducers and deleted unused animalsReducer
