+++ src/presentation/components/Navbar.tsx
@@ -13,7 +13,7 @@ import { AppRoutes } from '../../routes/app-routes'
 import { useLocation, useNavigate } from 'react-router-dom'
 import Logo from '../assets/Logo.png'
 import { useSelector, useDispatch } from 'react-redux'
-import { logout } from '../reducer/userReducer'
+import { logout } from '../reducer/user-reducer'
 import useNotify from '../hooks/use-notify'
 import { OngModel } from '../models/ong-model'
 import useAuth from '../hooks/use-auth'
