+++ src/presentation/pages/Login.tsx
@@ -14,7 +14,7 @@ import { EyeFilledIcon } from '../assets/EyeFilledIcon'
 import useNotify from '../hooks/use-notify'
 import { useFetch } from '../hooks/use-fetch'
 import { useDispatch } from 'react-redux'
-import { login } from '../reducer/userReducer'
+import { login } from '../reducer/user-reducer'
 import { useNavigate } from 'react-router'
 import useAuth from '../hooks/use-auth'
 import { OngInteractor } from '../../domain/interactors/ong-interactor'
