+++ src/presentation/store/index.ts
@@ -1,11 +1,8 @@
-import { configureStore } from '@reduxjs/toolkit';
-import userReducer from '../reducer/userReducer';
-import animalsReducer from '../reducer/animalsReducer';
+import { configureStore } from '@reduxjs/toolkit'
+import userReducer from '../reducer/user-reducer'
 
 export const store = configureStore({
     reducer: {
         user: userReducer,
-        animals: animalsReducer,
-    }
+    },
 })
-
